refactor: clean types, consts, enums
