+++ src/components/dashboard/types.ts
@@ -1,5 +1,1 @@
-export type ResumeListItem = {
-  id: string;
-  original_filename: string;
-  created_at: string | null;
-};
+export type { ResumeListItem } from "~/types/resume";
