+++ src/enums/auth.ts
@@ -1,0 +1,4 @@
+export enum AuthProviderEnum {
+  GITHUB = "github",
+  GOOGLE = "google",
+}
