+++ src/enums/resume.ts
@@ -1,0 +1,12 @@
+export enum ResumeFileTypeEnum {
+  PDF = "pdf",
+  DOCX = "docx",
+}
+
+export enum ResumeUploadStateEnum {
+  IDLE = "idle",
+  UPLOADING = "uploading",
+  SAVING = "saving",
+  SUCCESS = "success",
+  FAILED = "failed",
+}
