+++ src/enums/supabase.ts
@@ -1,0 +1,7 @@
+export enum SupabaseAuthKeyEnum {
+  AUTH = "auth",
+}
+
+export enum SupabaseAuthMethodKeyEnum {
+  GET_SESSION = "getSession",
+}
