+++ src/types/content.ts
@@ -1,0 +1,5 @@
+export type PostType = {
+  id: number;
+  title: string;
+  body: string;
+};
