+++ src/types/routeAuth.ts
@@ -1,0 +1,14 @@
+export type RouteLocationLike = {
+  pathname: string;
+  search: unknown;
+};
+
+export type RequireDashboardAuthOptions = {
+  location: RouteLocationLike;
+  hasKnownUser?: boolean;
+};
+
+export type RedirectAuthenticatedFromLoginOptions = {
+  redirectPath: string | undefined;
+  hasKnownUser?: boolean;
+};
