+++ src/types/user.ts
@@ -1,0 +1,5 @@
+export type User = {
+  id: number;
+  name: string;
+  email: string;
+};
