+++ src/utils/users.tsx
@@ -1,5 +1,1 @@
-export type User = {
-  id: number
-  name: string
-  email: string
-}
+export type { User } from "~/types/user";
